requirements/common.txt
flask==1.1.1
gunicorn==20.0.4
gevent==1.4.0
structlog==20.1.0
colorama==0.4.3
logstash-formatter==0.5.17
flask-sqlalchemy==2.4.1
psycopg2==2.8.4
requests==2.23.0
bcrypt==3.1.7
marshmallow==3.5.1
